18f6d043b81e6fa54ec6eb2aa0b4322d2a2fd842 [Trivial] Get rid of compiler warnings (random-zebra)

Pull request description:

  This removes a few compiler warnings:

  uninitialized variables:
  <code>
  init.cpp:1620:47: warning: ‘nStart’ may be used uninitialized in this function [-Wmaybe-uninitialized]
  </code>
  <strike>
  leveldb/port/port_posix.cc:60:15: warning: ‘ecx’ may be used uninitialized in this function [-Wmaybe-uninitialized]
  </strike>
  <br>

  signed/unsigned int comparison:
  <code>
  wallet/wallet.cpp:2446:39: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  </code>
  <code>
  utilstrencodings.cpp:58:21: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
  </code>
  <br>

  variables defined/set but not used
  <strike>
  qt/privacydialog.cpp:710:13: warning: variable ‘nLockedBalance’ set but not used [-Wunused-but-set-variable]
  </strike>
  <code>
  ./qt/pivx/qtutils.h:22:21: warning: ‘SHORT_KEY’ defined but not used [-Wunused-variable]
  </code>

ACKs for top commit:
  furszy:
    ACK 18f6d043b81e6fa54ec6eb2aa0b4322d2a2fd842
  Fuzzbawls:
    ACK 18f6d043b81e6fa54ec6eb2aa0b4322d2a2fd842

Tree-SHA512: 46830addd7a21b07e46a9eb53354d50b51d3846b590925e524f68b92cbc8cd71f8b29fca5ab8e399f300164e7f895d634ff794a6f55bfa097a7ec54a54084743

+++ src/utilstrencodings.cpp
@@ -55,7 +55,7 @@ bool validateURL(std::string strURL, std::string& strErr, unsigned int maxSize) 
 
     // check fronts
     bool found = false;
-    for (int i=0; i < reqPre.size() && !found; i++) {
+    for (int i=0; i < (int) reqPre.size() && !found; i++) {
         if (strURL.find(reqPre[i]) == 0) found = true;
     }
     if ((!found) && (reqPre.size() > 0)) {
